Deleted .DS_Store
Attributes files like this and several other files that are not necessary to the project when collaborating are supposed to be ignored by our .gitignore file we have setup. For some reason Eclipse is not honoring our .gitignore file in it's entirety. I've noticed a couple of others that are not being ignored either but they are listed in the .gitignore file.