redirect to index.jsp
The default site preferences of most web servers for the directory index file is index.html (Unless configured specifically to include index.jsp). Not knowing where a project might be deployed and what default configuration might be in place (or available to be modified) a simple index.html redirect would be the simplest solution.